refactor: studyRecordHandler 제거 및 queryHandler 추가 및 적용
